Move logic to base-element
