+++ extensions/amp-sidebar/1.0/base-element.js
@@ -17,8 +17,46 @@
 import {CSS as COMPONENT_CSS} from './component.jss';
 import {PreactBaseElement} from '../../../src/preact/base-element';
 import {Sidebar} from './component';
+import {dict} from '../../../src/utils/object';
+import {toggle} from '../../../src/style';
+import {toggleAttribute} from '../../../src/dom';
 
-export class BaseElement extends PreactBaseElement {}
+export class BaseElement extends PreactBaseElement {
+  /** @param {!AmpElement} element */
+  constructor(element) {
+    super(element);
+
+    /** @private {boolean} */
+    this.open_ = false;
+  }
+
+  /** @override */
+  init() {
+    return dict({
+      'onBeforeOpen': this.toggle_.bind(this, true),
+      'onAfterClose': this.toggle_.bind(this, false),
+    });
+  }
+
+  /**
+   * Toggle open/closed attributes.
+   * @param {boolean} opt_state
+   */
+  toggle_(opt_state) {
+    this.open_ = toggleAttribute(this.element, 'open', opt_state);
+    toggle(this.element, this.open_);
+  }
+
+  /** @override */
+  mutationObserverCallback() {
+    const open = this.element.hasAttribute('open');
+    if (open === this.open_) {
+      return;
+    }
+    this.open_ = open;
+    open ? this.api().open() : this.api().close();
+  }
+}
 
 /** @override */
 BaseElement['Component'] = Sidebar;
